dom/tests/mochitest/gamepad/test_gamepad_iframe_insecure.html

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test gamepad</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script type="text/javascript" src="mock_gamepad.js"></script>
<script class="testbody" type="text/javascript">
const { is, ok, SimpleTest, SpecialPowers } = window.parent;

window.addEventListener("gamepadconnected", () => {
  ok(false, "gamepadconnected event must not fire");
});

window.addEventListener("gamepaddisconnected", () => {
  ok(false, "gamepaddisconnected event must not fire");
});


runGamepadTest(async ()=> {
  // Add a gamepad - normally would cause "gamepadconnected" to fire
  const index = await GamepadService.addGamepad("test gamepad", // id
    GamepadService.standardMapping,
    GamepadService.noHand,
    4,
    2,
    0,
    0,
    0
  );
  is(navigator.getGamepads().length, 0, "Expected 0 gamepads in insecure context");
  await GamepadService.removeGamepad(index);
  SimpleTest.finish();
});
</script>
</body>
</html>
